requirements.txt
psycopg2-binary==2.9.5
google-cloud-texttospeech
requests
pandas
finsymbols
regex